Change temp variable type from int to double

+++ 02. Conditional Statements/12. Temperature .cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main ( )
 {
-    int temp;
+    double temp;
     cout<<"Enter Temperature:";
     cin>>temp;
